lint: clear arc-crypto and the workspace-wide auto-fixable clippy findings
`make lint` (clippy --workspace --all-targets -D warnings) has not been
passing. It stopped at arc-crypto, which alone had 26 errors, and arc-crypto
failing masked every downstream crate.

arc-crypto is now clean. 20 of its 26 were mechanical and machine-applied
(x % 2 == 0 -> is_multiple_of, (a + b - 1) / b -> div_ceil, acc = acc + x ->
acc += x, redundant closures, repeat().take() -> repeat_n). All are identity
transformations; arc-crypto's 230 tests pass unchanged before and after.

The six that needed hands:
  - IncrementalMerkle: added the Default impl to match new().
  - poseidon MDS mix: rewrote the range loop as mds.iter().enumerate(), using
    the row directly. `mds` is a local Vec from generate_mds_matrix, not a
    borrow of self, so this cannot alias self.elements.
  - Lagrange interpolation: destructured (x_j, _) in the enumerate pattern.
  - batch_prover rolling average: checked_div now carries the total_proofs > 0
    guard the explicit `if` was providing.
  - zk_rollup x2: &[tx.clone()] -> std::slice::from_ref(&tx).

Then ran the auto-fix across the whole workspace. That introduced one
regression which is fixed here: it stripped `Address` from block_stm.rs's
imports because only cfg(test) code uses it, breaking the arc-state test
build. Imported in the two test modules that need it rather than restored at
module scope, where the lib target would rightly call it unused.

Whole workspace: 1,372 lib tests pass, 0 failures.

93 clippy findings remain and are NOT addressed here. They are a different
class -- ~35 needless_range_loop in the inference/GPU matrix kernels, 20 float
strict-comparisons, 10 complex-type, 9 constant assertions. The matrix loops
sit on the INT16 determinism path, where a careless rewrite would silently
break the bitwise-identical-output property the chain's thesis rests on. They
need reviewing individually against the determinism tests, not machine-applied.

+++ crates/arc-inference/src/gguf_integer.rs
@@ -103,7 +103,7 @@ pub fn generate_integer_from_gguf(
     let n_kv_heads = get_meta_u32(&content, &format!("{arch}.attention.head_count_kv")).unwrap_or(n_heads as u32) as usize;
     let d_ff = get_meta_u32(&content, &format!("{arch}.feed_forward_length"))? as usize;
     let vocab_size = content.tensor_infos.get("token_embd.weight")
-        .map(|t| t.shape.dims()[0] as usize)
+        .map(|t| t.shape.dims()[0])
         .unwrap_or(32000);
 
     // Read RoPE base frequency from metadata (LLaMA-3 uses 500000, most others use 10000)
